converted Card component to be JSX based
