+++ .nvmrc
@@ -1,1 +1,1 @@
-18.12.1
+18.15.0
