+++ src/templates/app.html
@@ -9,7 +9,7 @@
 
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro">
     <link rel="stylesheet" href="../styles/main.css">
-    <script type="module" src="./components/card.js"></script>
+    <script type="module" src="../components/card.jsx?type=jsx"></script>
     <script src="/node_modules/htmx.org/dist/htmx.js"></script>
   </head>
